+++ tests/atest/transformers/RemoveEmptySettings/test_transformer.py
@@ -1,9 +1,9 @@
 import pytest
 
-from .. import run_tidy_and_compare
+from .. import TransformerAcceptanceTest
 
 
-class TestRemoveEmptySettings:
+class TestRemoveEmptySettings(TransformerAcceptanceTest):
     TRANSFORMER_NAME = "RemoveEmptySettings"
 
     @pytest.mark.parametrize("source", ["empty", "overwritten"])
@@ -12,8 +12,7 @@ class TestRemoveEmptySettings:
     def test_modes(self, source, work_mode, more_explicit):
         work_mode_config = "" if work_mode == "overwritten_ok" else f":work_mode={work_mode}"
         more_explicit_config = "" if more_explicit == "more_explicit" else ":more_explicit=False"
-        run_tidy_and_compare(
-            self.TRANSFORMER_NAME,
+        self.compare(
             source=f"{source}.robot",
             expected=f"{source}_{work_mode}_{more_explicit}.robot",
             config=f"{work_mode_config}{more_explicit_config}",