+++ tests/atest/transformers/SplitTooLongLine/test_transformer.py
@@ -1,36 +1,32 @@
-from .. import run_tidy_and_compare
+from .. import TransformerAcceptanceTest
 
 
-class TestSplitTooLongLine:
+class TestSplitTooLongLine(TransformerAcceptanceTest):
     TRANSFORMER_NAME = "SplitTooLongLine"
 
     def test_split_too_long_lines(self):
-        run_tidy_and_compare(
-            self.TRANSFORMER_NAME,
+        self.compare(
             source="tests.robot",
             expected="feed_until_line_length.robot",
             config=":line_length=80:split_on_every_arg=False -s 4",
         )
 
     def test_split_too_long_lines_split_on_every_arg(self):
-        run_tidy_and_compare(
-            self.TRANSFORMER_NAME,
+        self.compare(
             source="tests.robot",
             expected="split_on_every_arg.robot",
             config=":line_length=80:split_on_every_arg=True -s 4",
         )
 
     def test_split_lines_with_multiple_assignments(self):
-        run_tidy_and_compare(
-            self.TRANSFORMER_NAME,
+        self.compare(
             source="multiple_assignments.robot",
             expected="multiple_assignments_until_line_length.robot",
             config=":line_length=80:split_on_every_arg=False -s 4",
         )
 
     def test_split_lines_with_multiple_assignments_on_every_arg(self):
-        run_tidy_and_compare(
-            self.TRANSFORMER_NAME,
+        self.compare(
             source="multiple_assignments.robot",
             expected="multiple_assignments_on_every_arg.robot",
             config=":line_length=80:split_on_every_arg=True -s 4",